refactored exports & fixed typos
